Task #5077: MErge from trunk

+++ branches/Cobalt-Task5077/RTCP/Cobalt/GPUProc/src/cuda/Kernels/FIR_FilterKernel.h
@@ -44,22 +44,22 @@ namespace LOFAR
       struct Parameters : Kernel::Parameters
       {
         Parameters(const Parset& ps);
-        size_t nrBitsPerSample;
-        size_t nrBytesPerComplexSample;
+        unsigned nrBitsPerSample;
+        unsigned nrBytesPerComplexSample;
 
         // The number of stations or TABs to filter. The FIR filter will
         // deal with either in the same way.
-        size_t nrSTABs;
+        unsigned nrSTABs;
 
         // The number of subbands \e this kernel instance will process,
         // typically equal to \c nrSubbandsPerSubbandProc.
-        size_t nrSubbands;
+        unsigned nrSubbands;
 
         // The number of PPF filter taps.
-        static const size_t nrTaps = 16;
+        static const unsigned nrTaps = 16;
 
         // The number of history samples used for each block
-        size_t nrHistorySamples() const;
+        unsigned nrHistorySamples() const;
       };
 
       enum BufferType
@@ -91,12 +91,12 @@ namespace LOFAR
 
       void enqueue(const BlockID &blockId,
                    PerformanceCounter &counter,
-                   size_t subbandIdx);
+                   unsigned subbandIdx);
 
       // Put the historyFlags[subbandIdx] in front of the given inputFlags,
       // and update historyFlags[subbandIdx] with the flags of the last samples
       // in inputFlags.
-      void prefixHistoryFlags(MultiDimArray<SparseSet<unsigned>, 1> &inputFlags, size_t subbandIdx);
+      void prefixHistoryFlags(MultiDimArray<SparseSet<unsigned>, 1> &inputFlags, unsigned subbandIdx);
 
     private:
       // The Kernel parameters as given to the constructor
